requirements.txt
# Core dependencies
pdfplumber>=0.10.0
pypdf>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# NLP and ML
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0

# Web UI
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
python-magic>=0.4.27
